Improve the syntax renderer by reusing original text
Now when the syntax produced by rules is turned back into a string, if any subforms are `syntax-original?`, the syntax renderer simply copies the exact text of the syntax object as found in the original file. This preserves formatting for unchanged subforms in refactored code, though it still looks pretty messy and it doesn't make any indentation corrections.