Try to deflake CI
Recently, tests have been failing randomly and inconsistently with GDK errors. Working theory is that xvfb-run isn't cooperating well with the way `raco test --drdr` runs each test in a separate process and runs multiple tests in parallel. This change removes the `--drdr` mode. Note that `--package` still causes each test to run in a separate process, but (to my knowledge) does _not_ cause tests to run in parallel. So hopefully this removes the flakiness.