Allow specifying multiple inputs in refactoring tests
The test checks that *each input* produces the expected output (the last code block/line is always the expected output). This makes it easier to write tests for conditionals and comparisons, since those test cases tend to have a combinatoric explosion in the number of ways that the same code can be written.