Add support for only analyzing ranges of files
This commit introduces the ability for Resyntax to limit its analysis to only certain lines in a file. This doesn't yet hook things up to git to actually figure out what the changed ranges of files in a diff are, but it's a step. The functionality only exists deep in the internals of Resyntax. There's no CLI for specifying what ranges of a file to run Resyntax on. Next I have to implement some code that parses `git diff` output to figure out where the changed ranges in a file are.