Fix a bug in diff analysis
The current logic is buggy and fails for changes like [this](https://github.com/jackfirth/yaragg/pull/5). The correct behavior is for Resyntax to make sure that all of the changed lines in a suggestion are included in the set of lines changed by a pull request, rather than merely checking that these two sets merely intersect.