Add `~replacement` metafunction (#247)
This cooperates with `syntax-mark-original-neighbors` and `syntax-originally-neighbors?` to detect when two syntax objects weren't originally adjacent in the input syntax, but one (or both) of them is a drop-in replacements for input subforms that _were_ adjacent. This makes a lot of uses of `ORIGINAL-GAP` unnecessary, though sadly not all of them.