Fix a bug caused by `define-signature`
The `(open ...)` clause of `define-signature` _somehow_ introduces syntax objects in its expansion that are `syntax-original?`, but can be from a different file. This causes the linemap checking that Resyntax performs to break, since the line numbers from those syntax objects won't match up with the file being analyzed. This commit ensures that Resyntax only looks at original syntax that's from the same file as the one being analyzed.