Avoid error on analyzing non-racket files
When `resyntax-analyze` is called directly on a non-`#lang racket` file, it currently results in an error during macro expansion being raised and logged. This causes issues when the langserver tries to analyze the current file with Resyntax, see jeapostrophe/racket-langserver#153. This PR moves Resyntax's logic for handling such files out of the file group resolution code and into the core of `resyntax-analyze`.